Float stations are parsed correctly.
The entire string was being treated as a list of stations, including the `hfloat` and `vfloat` values. So this:
```adj.setConfig("float", "10.0 91.0 ST1")```

Was being parsed as three stations "10.0", "91.0", and "ST1" each with an `hfloat` and `vfloat` value of 10.0 and 91.0 respectively.

This fixes this bug by passing the `slist` value into `_splitStationList` rather than `value`.